AIService/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.4
opencv-python-headless>=4.9.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
